MySportsFeeds Seasonal Games

Returns a data frame of games from MySportsFeeds version 2.0 API

msf_seasonal_games(league, season, verbose = FALSE)

Arguments

league
the league to fetch

season
the season to fetch

verbose
print status info

Value

a list of two items

status_code the HTTP status code (200 for success)

response if status_code == 200, a tibble of games; otherwise, the raw text.

Details

msf_seasonal_games adds four columns at the right of the tibble:

league the source league of the data

season the source season of the data, and

date the game date (started) in the Eastern USA timezone ("EST5EDT").

slug the game slug (date-away_team-home_team)

The returned tibble will be sorted in chronological order.

Examples

# NOT RUN {
nba_games <- dfstools::msf_seasonal_games(
season = "2017-2018-regular", league = "nba"
)
# }